pyproject.toml
[tool.poetry]
name = "tukro"
version = "0.1.0"
description = ""
authors = ["Mateus Ferreira <mateusasferreira28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-guardian = "^2.4.0"
pytest = "^7.3.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
